Calculate fee based on entire payment transaction instead of individual payees
